+++ bin/tmctl.sh
@@ -589,7 +589,7 @@ cmd_server_edit() {
 }
 
 cmd_version() {
-    echo "TimeMachine Backup v3.0.2"
+    echo "TimeMachine Backup v3.0.3"
 }
 
 cmd_fix_permissions() {